refac: move argument handling to calculator_args.py (#1306)
