+++ source/databricks/calculation-engine/tests/integration/calculator/calculator_job/conftest.py
@@ -13,11 +13,10 @@
 # limitations under the License.
 
 from datetime import datetime
-from os import path
 from pyspark.sql import SparkSession, DataFrame
 import pytest
-from shutil import rmtree
 from typing import Callable, Optional
+from unittest.mock import patch, Mock
 
 from . import configuration as C
 from package.calculator_job import (
@@ -87,7 +86,7 @@ def executed_calculation_job(
         mode="overwrite",
     )
 
-    _start_calculator(spark, test_data_job_parameters)
+    _start_calculator(test_data_job_parameters, spark)
 
 
 @pytest.fixture(scope="session")
